refactor: modified folder structure Next v13
