+++ src/app/page.tsx
@@ -1,5 +1,5 @@
 import Image from "next/image";
-import Login from "../pages/login/";
+import Login from "../app/login/page";
 
 export default function Home() {
   return <Login />;
